+++ .gitignore
@@ -1,2 +1,4 @@
 .idea
 **.iml
+__pycache__
+**.egg-info
