--- requirements.txt
@@ -1,2 +1,0 @@
-pip3 install python-swiftclient
-pip3 install pymongo
